api/requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
multidict==6.0.4
oauthlib==3.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0
yarl==1.9.4
